+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.9.x
 
-- [**0.9.8.6**](#0986)
+- [**0.9.8.7**](#0987)
+- [0.9.8.6](#0986)
 - [0.9.8.5](#0985)
 - [0.9.8.4](#0984)
 - [0.9.8.3](#0982)
@@ -99,6 +100,28 @@
 ---
 
 # 0.9.x
+
+## 0.9.8.7
+
+### Anvil
+
+- The `Fonts` object will now create the `default8` texture in the resource pack instead of `/assets`
+- Update format version to `1.26.34`.
+- Increased the file path length from 80 to 96.
+
+### Actors
+
+- You can now customize your initial state name in client/server animation controller with "default" as the _default_ name.
+- Server animations now require a list of commands/events instead of a star-args.
+- Creating a render controller with the same name will now return the existing controller if it exists instead of overwriting it.
+
+### Blocks
+
+- Added `minecraft:sixteen_way_rotation` trait.
+
+### Molang
+
+- Renamed `Molang.molang_condition()` to `Molang.condition()`
 
 ## 0.9.8.6